Backend/lambdas/trade_ingest/requirements.txt
aws-msk-iam-sasl-signer-python
confluent-kafka